#66 REFACTOR open_api 패키지 -> openApi
