+++ src/main/java/com/sparta/topster/domain/openApi/service/maniadb/ManiadbService.java
@@ -1,12 +1,11 @@
-package com.sparta.topster.domain.open_api.service.maniadb;
+package com.sparta.topster.domain.openApi.service.maniadb;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
 import com.sparta.topster.domain.album.entity.Album;
-import com.sparta.topster.domain.open_api.service.OpenApiService;
+import com.sparta.topster.domain.openApi.service.OpenApiService;
 import com.sparta.topster.domain.song.entity.Song;
-import com.sparta.topster.domain.song.repository.SongRepository;
 import com.sparta.topster.global.exception.ServiceException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -14,7 +13,6 @@ import org.json.JSONArray;
 import org.json.JSONObject;
 import org.json.XML;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.context.annotation.Primary;
 import org.springframework.http.RequestEntity;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
@@ -26,7 +24,7 @@ import java.net.URI;
 import java.util.ArrayList;
 import java.util.List;
 
-import static com.sparta.topster.domain.open_api.exception.ManiadbException.NOT_SERCH_ALBUM;
+import static com.sparta.topster.domain.openApi.exception.ManiadbException.NOT_SERCH_ALBUM;
 
 @Service
 @Qualifier("maniadb")
