+++ src/main/java/com/sparta/topster/domain/openApi/service/spotify/controller/SpotifyController.java
@@ -1,9 +1,8 @@
-package com.sparta.topster.domain.open_api.service.spotify.controller;
+package com.sparta.topster.domain.openApi.service.spotify.controller;
 
-import com.sparta.topster.domain.open_api.service.spotify.SpotifyService;
+import com.sparta.topster.domain.openApi.service.spotify.SpotifyService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
